More README edits, renamed app to UsageRuleTUISample
