+++ UsageRuleSample/app/proguard-rules.pro
