+++ UsageRuleSample/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".MainActivity">
+    tools:context="com.rivetz.UsageRuleTUISample.MainActivity">
 
     <RelativeLayout
         android:layout_width="fill_parent"
