+++ UsageRuleSample/build.gradle
