+++ UsageRuleSample/gradle.properties
