+++ UsageRuleSample/gradle/wrapper/gradle-wrapper.jar
