+++ UsageRuleSample/gradle/wrapper/gradle-wrapper.properties
