+++ UsageRuleSample/gradlew.bat
