+++ UsageRuleSample/settings.gradle
